🚚 ♻️ refactored code and moved browser pool
